Added StarTeamAutoLabel for facilitating repository versioning. See docs for details.
Label functionality refactored into baseclass (LabelTask.)

StarTeamCheckin : If label property is set. After checking file in it movesthe specified label to the new revision of the file.

Output : If a file is checked out or in it is Logged. Otherwise a only tally of files affected unless verbose is turned on.