Added warning to msi if ProductName, ProductVersion, ProductCode, and Manufacturer are not specified properties
Removed requirement of ProductName, ProductVersion, ProductCode, and Manufacturer from the msm task

FeatureComponents table is not populated for the msm task

license, banner, and background attributes are now only part of the msi task.

Fixed the registry bug for obtaining/specifying the default value of a registry key.

Fixed the bug for the temporary directory existing if a file has the read-only attribute assigned.  The default behavior is to remove the read-only attribute from all files added to a cab file.  You can specify read-only attributes for the file(s) via attributes from the component or file override elements.