Removed ComponentCount index as it was not used.
Added support for installing assemblies to the GAC

Support for registry components

Added dependent registry root.  If it is a per-user install, the registry value is written under HKCU;
if it is a per-machine install, the value is written under HKLM.  Set to a per-machine by setting ALLUSERS
property to 1.